+++ eras/mary/impl/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### `testlib`
 
+* Move `mkTokenMintingTx` to `cardano-ledger-alonzo:testlib`
 * Add `DecCBOR` instance for `Block`
 
 ### `cddl`